Renamed some packages in the spring_data adapter
- Make some names shorter, making them equal to the rest of the services
