+++ users-service-adapters/users-service-spring-data-adapter/src/main/java/ar/edu/itba/cep/users_service/spring_data/SpringDataUserCredentialRepositoryAdapter.java
@@ -1,9 +1,9 @@
-package ar.edu.itba.cep.users_service.spring_data_repositories;
+package ar.edu.itba.cep.users_service.spring_data;
 
 import ar.edu.itba.cep.users_service.models.User;
 import ar.edu.itba.cep.users_service.models.UserCredential;
 import ar.edu.itba.cep.users_service.repositories.UserCredentialRepository;
-import ar.edu.itba.cep.users_service.spring_data_repositories.spring_data_interfaces.SpringDataUserCredentialRepository;
+import ar.edu.itba.cep.users_service.spring_data.interfaces.SpringDataUserCredentialRepository;
 import com.bellotapps.webapps_commons.persistence.spring_data.repository_utils_adapters.repositories.BasicRepositoryAdapter;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
