+++ users-service-adapters/users-service-spring-data-adapter/src/main/java/ar/edu/itba/cep/users_service/spring_data/config/SpringDataConfig.java
@@ -1,4 +1,4 @@
-package ar.edu.itba.cep.users_service.spring_data_repositories.config;
+package ar.edu.itba.cep.users_service.spring_data.config;
 
 import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.context.annotation.ComponentScan;
@@ -10,10 +10,10 @@ import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
  */
 @Configuration
 @ComponentScan(basePackages = {
-        "ar.edu.itba.cep.users_service.spring_data_repositories"
+        "ar.edu.itba.cep.users_service.spring_data"
 })
 @EnableJpaRepositories(basePackages = {
-        "ar.edu.itba.cep.users_service.spring_data_repositories.spring_data_interfaces"
+        "ar.edu.itba.cep.users_service.spring_data.interfaces"
 })
 @EntityScan(basePackages = {
         "ar.edu.itba.cep.users_service.models"
